refactor: 컨벤션 정리 — 깊은 경로 import·raw URL·목 데이터 위치
- PencilIcon/ProfileImageIcon: cn import를 @/lib/utils → @/lib 최상위 경로로 통일
- PATH_NAME.mypage.main이 tab 인자를 받도록 확장, Container의 수동 쿼리 조립(buildMypageHref) 제거
- BookDetail 인라인 목 데이터를 lib/mocks(MOCK_BOOK_SUMMARIES)로 이동
