+++ src/components/common/Icon/PencilIcon.tsx
@@ -1,6 +1,6 @@
 import { type SVGProps } from 'react';
 import { Pencil } from '@/assets';
-import { cn } from '@/lib/utils';
+import { cn } from '@/lib';
 
 type Props = SVGProps<SVGSVGElement>;
